feat(#69): 전체적인 구조 수정
UserController에서 사용자 인증과 관련된 회원가입, 로그인, 로그아웃, 회원탈퇴 api 를 AuthController에 나눠서 작성
jwt 관련 기능을 더 세부적으로 나누기
