--- server/src/main/java/com/hexacore/tayo/auth/RefreshTokenRepository.java
@@ -1,10 +1,0 @@
-package com.hexacore.tayo.auth;
-
-import com.hexacore.tayo.auth.model.RefreshTokenEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RefreshTokenRepository extends JpaRepository<RefreshTokenEntity, Long> {
-
-}
