+++ server/src/main/java/com/hexacore/tayo/auth/interceptor/AuthenticationInterceptor.java
@@ -1,8 +1,9 @@
 package com.hexacore.tayo.auth.interceptor;
 
+import com.hexacore.tayo.auth.jwt.JwtParser;
+import com.hexacore.tayo.util.RequestParser;
 import com.hexacore.tayo.common.errors.AuthException;
 import com.hexacore.tayo.common.errors.ErrorCode;
-import com.hexacore.tayo.auth.JwtService;
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -15,7 +16,10 @@ import java.util.Date;
 
 @RequiredArgsConstructor
 public class AuthenticationInterceptor implements HandlerInterceptor {
-    private final JwtService jwtService;
+    private final JwtParser jwtParser;
+
+    @Value("${jwt.secret-key}")
+    private String secretKey;
 
     @Value("${jwt.access.cookie-name}")
     private String accessTokenCookieName;
@@ -29,13 +33,12 @@ public class AuthenticationInterceptor implements HandlerInterceptor {
      */
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-
         if (handler.getClass().equals(ResourceHttpRequestHandler.class)) {
             return true;
         }
 
         String accessToken = RequestParser.getToken(request, accessTokenCookieName);
-        Claims claims = jwtService.getClaims(accessToken);
+        Claims claims = jwtParser.getClaims(accessToken);
 
         // 엑세스 토큰의 만료여부를 확인
         if (claims.getExpiration().after(new Date())) { // 만료되지 않은 토큰인 경우
