+++ server/src/main/java/com/hexacore/tayo/auth/interceptor/RefreshAccessTokenInterceptor.java
@@ -1,8 +1,9 @@
 package com.hexacore.tayo.auth.interceptor;
 
+import com.hexacore.tayo.auth.jwt.JwtParser;
+import com.hexacore.tayo.util.RequestParser;
 import com.hexacore.tayo.common.errors.AuthException;
 import com.hexacore.tayo.common.errors.ErrorCode;
-import com.hexacore.tayo.auth.JwtService;
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -14,7 +15,7 @@ import java.util.Date;
 
 @RequiredArgsConstructor
 public class RefreshAccessTokenInterceptor implements HandlerInterceptor {
-    private final JwtService jwtService;
+    private final JwtParser jwtParser;
 
     @Value("${jwt.refresh.cookie-name}")
     private String refreshTokenCookieName;
@@ -23,7 +24,7 @@ public class RefreshAccessTokenInterceptor implements HandlerInterceptor {
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
 
         String clientRefreshToken = RequestParser.getToken(request, refreshTokenCookieName);
-        Claims clientClaims = jwtService.getClaims(clientRefreshToken);
+        Claims clientClaims = jwtParser.getClaims(clientRefreshToken);
 
         // 리프레시 토큰 만료 여부를 확인
         if (clientClaims.getExpiration().before(new Date())) {
@@ -31,8 +32,7 @@ public class RefreshAccessTokenInterceptor implements HandlerInterceptor {
         }
 
         Long userId = Long.valueOf(clientClaims.getSubject());
-        String serverRefreshToken = jwtService.getRefreshToken(userId);
-        jwtService.checkTokenValidation(serverRefreshToken);
+        String serverRefreshToken = jwtParser.getValidRefreshToken(userId);
 
         // 클라이언트로부터 받은 리프레시 토큰과, 디비에 저장해 놓은 리프레시 토큰 값이 같은지 비교
         if (clientRefreshToken.equals(serverRefreshToken)) {
