+++ server/src/main/java/com/hexacore/tayo/config/InterceptorConfig.java
@@ -2,7 +2,7 @@ package com.hexacore.tayo.config;
 
 import com.hexacore.tayo.auth.interceptor.AuthenticationInterceptor;
 import com.hexacore.tayo.auth.interceptor.RefreshAccessTokenInterceptor;
-import com.hexacore.tayo.auth.JwtService;
+import com.hexacore.tayo.auth.jwt.JwtParser;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -12,28 +12,28 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 @Configuration
 @RequiredArgsConstructor
 public class InterceptorConfig implements WebMvcConfigurer {
-    private final JwtService jwtService;
+    private final JwtParser jwtParser;
 
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
         registry.addInterceptor(authenticationInterceptor())
                 .order(1)
                 .addPathPatterns("/**")
-                .excludePathPatterns("/signup", "/login", "/refresh"); // 로그인 인증 필요없는 요청
+                .excludePathPatterns("/auth/signup", "/auth/login", "/auth/refresh"); // 로그인 인증 필요없는 요청
 
         // 엑세스 토큰 재발급 요청인 경우
         registry.addInterceptor(refreshAccessTokenInterceptor())
                 .order(2)
-                .addPathPatterns("/refresh");
+                .addPathPatterns("/auth/refresh");
     }
 
     @Bean
     public RefreshAccessTokenInterceptor refreshAccessTokenInterceptor() {
-        return new RefreshAccessTokenInterceptor(jwtService);
+        return new RefreshAccessTokenInterceptor(jwtParser);
     }
 
     @Bean
     public AuthenticationInterceptor authenticationInterceptor() {
-        return new AuthenticationInterceptor(jwtService);
+        return new AuthenticationInterceptor(jwtParser);
     }
 }
