+++ server/src/main/java/com/hexacore/tayo/user/UserService.java
@@ -1,18 +1,14 @@
 package com.hexacore.tayo.user;
 
-import com.hexacore.tayo.auth.JwtService;
-import com.hexacore.tayo.car.model.CarEntity;
-import com.hexacore.tayo.car.model.ImageEntity;
+import com.hexacore.tayo.util.Encryptor;
 import com.hexacore.tayo.common.ResponseDto;
 import com.hexacore.tayo.common.errors.GeneralException;
 import com.hexacore.tayo.image.S3Manager;
 import com.hexacore.tayo.user.dto.*;
-import com.hexacore.tayo.common.errors.AuthException;
 import com.hexacore.tayo.common.errors.ErrorCode;
 import com.hexacore.tayo.user.model.UserEntity;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.mindrot.jbcrypt.BCrypt;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
@@ -20,43 +16,8 @@ import org.springframework.stereotype.Service;
 @Service
 @RequiredArgsConstructor
 public class UserService {
-    private final JwtService jwtService;
     private final UserRepository userRepository;
     private final S3Manager s3Manager;
-
-    @Transactional
-    public ResponseDto signUp(SignUpRequestDto signUpRequestDto) {
-        // todo 회원탈퇴했던 사용자가 다시 회원가입한 경우 고려해보기
-
-        // 이메일 중복 확인
-        if(userRepository.findByEmail(signUpRequestDto.getEmail()).isPresent()) {
-            throw new GeneralException(ErrorCode.USER_EMAIL_DUPLICATED);
-        }
-
-        // 프로필 사진 업로드
-        String profileUrl = null;
-        if (signUpRequestDto.getProfileImg() != null && !signUpRequestDto.getProfileImg().isEmpty()) {
-            profileUrl = s3Manager.uploadImage(signUpRequestDto.getProfileImg());
-        }
-
-        UserEntity newUser = UserEntity.builder()
-                .email(signUpRequestDto.getEmail())
-                .name(signUpRequestDto.getName())
-                .password(encryptPwd(signUpRequestDto.getPassword()))
-                .phoneNumber(signUpRequestDto.getPhoneNumber())
-                .profileImgUrl(profileUrl)
-                .build();
-
-        userRepository.save(newUser);
-        return ResponseDto.success(HttpStatus.CREATED);
-    }
-
-    public UserInfoResponseDto getUser(Long userId) {
-        UserEntity user = userRepository.findById(userId).orElseThrow(() ->
-                new GeneralException(ErrorCode.USER_NOT_FOUND));
-
-        return getUserInfo(user);
-    }
 
     @Transactional
     public ResponseDto update(Long userId, UserUpdateRequestDto updateRequestDto) {
@@ -72,7 +33,7 @@ public class UserService {
 
         // 새로운 비밀번호를 입력한 경우
         if (updateRequestDto.getPassword()!= null && !updateRequestDto.getPassword().isEmpty()) {
-            user.setPassword(encryptPwd(updateRequestDto.getPassword()));
+            user.setPassword(Encryptor.encryptPwd(updateRequestDto.getPassword()));
         }
 
         // todo null 체크를 해줘야할지 클라이언트 상에서 확인 후 수정
@@ -81,60 +42,11 @@ public class UserService {
         return ResponseDto.success(HttpStatus.OK);
     }
 
-    @Transactional
-    public void logOut(Long userId) {
-        jwtService.deleteRefreshToken(userId);
-    }
-
-    @Transactional
-    public void delete(Long userId) {
+    public UserInfoResponseDto getUser(Long userId) {
         UserEntity user = userRepository.findById(userId).orElseThrow(() ->
                 new GeneralException(ErrorCode.USER_NOT_FOUND));
 
-        // 유저 soft delete
-        user.setDeleted(true);
-
-        // 유저가 등록한 차가 있는 경우, 차도 soft delete
-        CarEntity userCar = user.getCar();
-        if (userCar != null) {
-            userCar.setIsDeleted(true);
-
-            // 차의 image 들 soft delete
-            for (ImageEntity carImage : userCar.getImages()) {
-                carImage.setIsDeleted(true);
-            }
-        }
-
-        // 발급받은 리프레시 토큰 삭제
-        jwtService.deleteRefreshToken(userId);
-    }
-
-    public LoginResponseDto login(LoginRequestDto loginRequestDto) {
-        UserEntity loginUser = userRepository.findByEmail(loginRequestDto.getEmail()).orElseThrow(() ->
-                new AuthException(ErrorCode.USER_NOT_FOUND));
-
-        if (loginUser.isDeleted()) {
-            throw new AuthException(ErrorCode.USER_DELETED);
-        }
-
-        if (!BCrypt.checkpw(loginRequestDto.getPassword(), loginUser.getPassword())) {
-            throw new AuthException(ErrorCode.USER_WRONG_PASSWORD);
-        }
-
-        UserInfoResponseDto loginUserInfo = getUserInfo(loginUser);
-
-        return LoginResponseDto.builder()
-                .accessToken(jwtService.createAccessToken(loginUser))
-                .refreshToken(jwtService.createRefreshToken(loginUser.getId()))
-                .loginUserInfo(loginUserInfo)
-                .build();
-    }
-
-    public String refresh(Long userId) {
-        UserEntity expiredUser = userRepository.findById(userId).orElseThrow(() ->
-                new AuthException(ErrorCode.USER_NOT_FOUND));
-
-        return jwtService.createAccessToken(expiredUser);
+        return getUserInfo(user);
     }
 
     private UserInfoResponseDto getUserInfo(UserEntity user) {
@@ -145,11 +57,5 @@ public class UserService {
                 .phoneNumber(user.getPhoneNumber())
                 .profileImgUrl(user.getProfileImgUrl())
                 .build();
-    }
-
-    // 비밀번호 암호화
-    private String encryptPwd(String plainPwd) {
-        String salt = BCrypt.gensalt();
-        return BCrypt.hashpw(plainPwd, salt);
     }
 }
